make basic styling responsive
