+++ src/components/footer/Footer.js
@@ -1,13 +1,20 @@
 import React, { Component } from 'react';
-import './footer.scss'
+import Header from '../header/Header';
 
 class componentName extends Component {
   render() {
     return (
       <footer className="footer">
-        <div className="container">
-          <p> Open Referral UK Data Standard maintained by Open Referral UK</p>
-          <p> Published with GitHub Pages </p>
+        <Header />
+        <div className="parent">   
+          <div>&copy; 2019-2021 Open Referral UK</div>
+          <div className="oruktools">
+            <ul>
+              <li> Terms &amp; Conditions</li>
+              <li>Pivacy Policy</li>
+            </ul>
+            </div>
+        
         </div>
       </footer>
     );
