+++ src/styles/css/card.css
@@ -1,0 +1,19 @@
+.card {
+  background-color: #f5f5f5;
+  flex: 0 3 24%;
+  border: 2px solid #e7e7e7;
+  border-radius: 8px;
+  padding: 1rem;
+  margin-bottom: 1rem;
+}
+
+.card img {
+  width: 100%;
+  object-fit: contain;
+}
+
+.card .date {
+  font-weight: bold;
+  text-align: right;
+}
+/*# sourceMappingURL=card.css.map */
