+++ src/styles/css/footer.css
@@ -1,0 +1,13 @@
+footer {
+  background-color: #212121;
+  color: white;
+  padding: 2rem 0;
+}
+
+footer .container {
+  margin: 0 auto;
+  width: 1080px;
+  display: flex;
+  flex-direction: column;
+}
+/*# sourceMappingURL=footer.css.map */
