+++ src/styles/sass/_md.scss
@@ -1,0 +1,56 @@
+.header {
+    grid-column-start: 2;
+    grid-column-end: 5;
+    grid-row-start: 1;
+    grid-row-end: 1;
+    background-color: $primaryBackgroudColour;
+  }
+
+  .nav {
+    grid-column-start: 2;
+    grid-column-end: 5;
+    grid-row-start: 2;
+    grid-row-end: 2;
+    background-color: $primaryBackgroudColour;
+  }
+
+  nav {
+
+    ul { 
+        display: flex;
+        justify-content: space-between;
+    }
+  
+    li {
+      margin-right: 15px;
+      list-style-type: none;
+    }
+  
+      ul li a {
+          color:$fontColor;
+      }
+}
+
+.sidebar {
+    grid-column-start: 2;
+    grid-column-end: 3;
+    grid-row-start: 4;
+    grid-row-end: 6;
+  }
+
+  .content {
+    grid-column-start: 3;
+    grid-column-end: 5;
+    grid-row-start: 4;
+    grid-row-end: 6;
+    min-height: 400px;
+  }
+
+  .footer {
+    grid-column-start: 2;
+    grid-column-end: 5;
+    grid-row-start: 6;
+    grid-row-end: 6;
+    background-color: $primaryBackgroudColour;
+	color: $fontColour;
+  }
