+++ src/styles/sass/_sm-default.scss
@@ -1,0 +1,98 @@
+/* Prevent adjustments of font size after orientation changes in IE on Windows Phone and in iOS.  */
+html {-webkit-text-size-adjust: 100%; -ms-text-size-adjust: 100%;}
+
+/*  apply a natural box layout model to all elements  */
+* { -moz-box-sizing: border-box; -webkit-box-sizing: border-box; box-sizing: border-box; }
+
+/* ------------- BODY ------------- */
+body {
+	font-family: $defaultFont;
+	font-size: 16px;
+	font-weight: $normalWeight;
+}
+
+.container {
+    display: grid;
+    grid-template-columns: 0.2fr 0.4fr 0.3fr 0.3fr 0.2fr;
+    grid-column-gap: 10px;
+    grid-row-gap: 15px;
+  }
+
+  .header {
+    grid-column-start: 1;
+    grid-column-end: 6;
+    grid-row-start: 1;
+    grid-row-end: 2;
+    background-color: $primaryColour;
+    color: $secondaryColour;
+    
+    a{
+        color:$fontColor;
+    }
+  }
+
+  nav {
+	
+	
+	ul { display: flex;}
+	ul li {
+		flex-basis: 16.66%;
+
+	}// end li
+	ul li a {
+		display: block;
+		text-align: center;
+		font-weight: $normalWeight;
+		padding: .5rem 0;
+		color:$fontColor;
+		text-decoration: none;
+	} 
+	ul li a:hover {
+		color: $secondaryColor;
+		background-color: $primaryColor;
+	} // end hover
+} // end nav
+
+  .sidebar {
+    grid-column-start: 2;
+    grid-column-end: 3;
+    grid-row-start: 4;
+    grid-row-end: 6;
+  }
+
+  .footer {
+    grid-column-start: 1;
+    grid-column-end: 6;
+    grid-row-start: 6;
+    grid-row-end: 6;
+    background-color: $primaryColour;
+	color: $secondaryColour;
+  }
+
+  footer ul { 
+    display: flex;
+    justify-content: space-between;
+}
+
+footer ul li {
+  margin-right: 15px;
+  list-style-type: none;
+}
+  
+  .content {
+    grid-column-start:1;
+    grid-column-end: 5;
+    grid-row-start: 3;
+    grid-row-end: 6;
+    height: 400px;
+  }
+ 
+  .parent {
+    display: flex;
+    padding: 0.5rem 0;
+  }
+
+  .oruktools {
+    margin-left: auto;
+    order: 2;
+  }
