+++ src/styles/sass/styles.scss
@@ -1,0 +1,13 @@
+@import "reset";
+
+@import "variables";
+
+@import "sm-default";
+
+@media only screen and (min-width: 35em) {
+	@import "md";
+}
+
+@media only screen and (min-width: 64em) {
+	@import "lg";
+}
